requirements.txt
blinker==1.9.0
click==8.1.8
Flask==3.1.2
importlib_metadata==8.7.0
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.3
python-dotenv==1.2.1
Werkzeug==3.1.3
zipp==3.23.0
